Adds endpoint for dice simulation list Using projection and distict to get the distinct value pair of dice number and side number out of the db

+++ src/main/java/com/example/avaloq/challenge/model/DiceSimulation.java
@@ -5,6 +5,6 @@ import lombok.Data;
 @Data
 public class DiceSimulation extends Dice{
 
-    private int totalNumberOfRolls;
-    private int simulationNumber;
+    private long totalNumberOfRolls;
+    private long simulationNumber;
 }
